add comprehensive numeric type support
- Signed integers: Int8, Int16, Int32, Int64
- Unsigned integers: UInt8, UInt16, UInt32, UInt64
- Floats: Float32, Float64
- Arbitrary precision: BigInt (using num-bigint)
- Fixed-point: Decimal (using rust_decimal)

Literal suffixes: 42i32, 100u16, 3.14f32, 42n, 99d

Conversion builtins: toInt8, toInt16, toInt32, toInt64,
toUInt8, toUInt16, toUInt32, toUInt64, toFloat32, toFloat64,
toBigInt

JIT extended to support all numeric types with type detection.
Polymorphic arithmetic dispatch at runtime.

+++ crates/vm/src/process.rs
@@ -320,14 +320,14 @@ mod tests {
         let mut receiver = Process::new(Pid(2));
 
         // Allocate a message on sender's heap
-        let msg = GcValue::Int(42);
+        let msg = GcValue::Int64(42);
 
         // Deliver to receiver (deep copy)
         receiver.deliver_message(msg, &sender.heap);
 
         assert!(receiver.has_messages());
         let received = receiver.try_receive().unwrap();
-        assert_eq!(received, GcValue::Int(42));
+        assert_eq!(received, GcValue::Int64(42));
         assert!(!receiver.has_messages());
     }
 
@@ -340,7 +340,7 @@ mod tests {
         assert_eq!(proc.state, ProcessState::Waiting);
 
         // Message delivery wakes up
-        proc.deliver_message(GcValue::Int(1), &Heap::new());
+        proc.deliver_message(GcValue::Int64(1), &Heap::new());
         assert_eq!(proc.state, ProcessState::Running);
     }
